moved StarRating files to shared folder
